refactor: rename settings_api package to settings
Renames the `settings_api` package to `settings` to improve clarity.
Also updates all imports to reflect the change.
