+++ app/lib/src/logic/composition_root.dart
@@ -2,7 +2,7 @@ import 'package:clock/clock.dart';
 import 'package:error_reporter/error_reporter.dart';
 import 'package:logger/logger.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:settings_api/settings_api.dart';
+import 'package:settings/settings.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:sizzle_starter/src/model/application_config.dart';
 import 'package:sizzle_starter/src/model/dependencies_container.dart';
