+++ app/lib/src/model/dependencies_container.dart
@@ -1,7 +1,7 @@
 import 'package:error_reporter/error_reporter.dart';
 import 'package:logger/logger.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:settings_api/settings_api.dart';
+import 'package:settings/settings.dart';
 import 'package:sizzle_starter/src/model/application_config.dart';
 
 /// Container for global dependencies.
