+++ app/lib/src/widget/dependencies_scope.dart
@@ -2,7 +2,7 @@ import 'package:common/common.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:settings_api/settings_api.dart';
+import 'package:settings/settings.dart';
 import 'package:sizzle_starter/src/model/dependencies_container.dart';
 
 class DependenciesScope extends StatelessWidget {
