+++ app/pubspec.yaml
@@ -40,8 +40,8 @@ dependencies:
     path: ../core/ui_library
 
   # Features
-  settings_api:
-    path: ../feature/settings/settings_api
+  settings:
+    path: ../feature/settings
   home:
     path: ../feature/home
