+++ feature/settings/lib/src/domain/model/general_configuration.dart
