+++ feature/settings/lib/src/injection.dart
@@ -1,6 +1,6 @@
-import 'package:settings_api/settings_api.dart';
-import 'package:settings_api/src/data/datasources/settings_local_datasource.dart';
-import 'package:settings_api/src/data/repositories/settings_repository_impl.dart';
+import 'package:settings/settings.dart';
+import 'package:settings/src/data/datasources/settings_local_datasource.dart';
+import 'package:settings/src/data/repositories/settings_repository_impl.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 Future<SettingsRepository> _createSettingsRepository(SharedPreferencesAsync sharedPreferences) {
